requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy-extensions==0.4.3
openai==0.26.1
pathspec==0.10.3
platformdirs==2.6.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.2
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
